11.10

+++ FamilyHealthCalendar/src/main/java/com/family/calendar/mapper/FamilyMemberMapper.java
@@ -7,7 +7,7 @@ import com.family.calendar.dtos.FamilyMemberDto;
 @Mapper
 public interface FamilyMemberMapper {
 
-    List<FamilyMemberDto> getMembersByUser(Long user_id);
+    List<FamilyMemberDto> getMembersByUser(int user_id);
 
     FamilyMemberDto getMemberDetail(Long member_id);
